# Conflicts:
#	bird/gugu-model/src/main/java/com/gugu/gugumodel/entity/strategy/MemberLimitStrategy.java
#	bird/gugu-model/src/main/java/com/gugu/gugumodel/mapper/StrategyMapper.java
#	bird/gugu-model/src/main/resources/xml/StrategyMapper.xml

+++ bird/gugu-model/src/main/java/com/gugu/gugumodel/dao/TeamDao.java
@@ -161,7 +161,7 @@ public class TeamDao{
      * 获取当前所有的serial
      * @param klassId
      */
-    public ArrayList<Integer> getSerial(Long klassId){
+    public ArrayList<Byte> getSerial(Long klassId){
         return teamMapper.getSerial(klassId);
     }
